Fixes issue #19 - Show associations as tree
Add a new pywbem instance subcommand, displaying a shrub where
a shrub is a view of the complete association relationship of a
particular cim instance showing not only the reference and
association information but the path of Role, AssocClass, ResultClass,
and ResultRole that formed the chain from the source instance
to the Result instances. It can display the complete set of
relationships or be limited with parameters that are filters to
particular sets of roles, association classes, and result roles.

The result by default is displayed as a tree in the following form

TST_Person.name="Mike"
 +-- member
 |   +-- TST_MemberOfFamilyCollection
 |       +-- family
 |           +-- TST_FamilyCollection(1 insts)
 |               +-- TST_FamilyCollection.name="Family2"
 +-- parent
     +-- TST_Lineage
         +-- child
             +-- TST_Person(2 insts)
                 +-- TST_Person.name="Sofi"
                 +-- TST_Person.name="Gabi"

The shrub may also display a table fo the results or the set
of objects that make up the result (MOF, etc.)

New Class

Creates a new file and new class, AssociationShrub that encapsulates all
of the code to gather data and present the shrub as either a table or an
ascii tree.